+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.6.0
+- Add option to output timestamp
+- Add option to disable color
+- Add `LogOutput`
+- Behaviour change of `LogPrinter`
+- Remove dependency
+
 ## 0.5.0
 - Add emojis
 - `LogFilter` is a class now